elixir json-ast: add AST structs and update golden
