+++ tools/json-ast/x/elixir/inspect.go
@@ -8,20 +8,6 @@ import (
 	ts "github.com/smacker/go-tree-sitter/elixir"
 )
 
-// Node represents a tree-sitter node.
-type Node struct {
-	Type     string `json:"type"`
-	Start    int    `json:"start"`
-	End      int    `json:"end"`
-	Text     string `json:"text,omitempty"`
-	Children []Node `json:"children,omitempty"`
-}
-
-// Program is the root of a parsed Elixir source file.
-type Program struct {
-	Root Node `json:"root"`
-}
-
 // Inspect parses Elixir source code using tree-sitter and returns the Program.
 func Inspect(src string) (*Program, error) {
 	parser := sitter.NewParser()
@@ -30,21 +16,6 @@ func Inspect(src string) (*Program, error) {
 	if err != nil {
 		return nil, fmt.Errorf("parse: %w", err)
 	}
-	root := convertNode(tree.RootNode(), []byte(src))
+	root := convert(tree.RootNode(), []byte(src))
 	return &Program{Root: root}, nil
-}
-
-func convertNode(n *sitter.Node, src []byte) Node {
-	node := Node{Type: n.Type(), Start: int(n.StartByte()), End: int(n.EndByte())}
-	if n.ChildCount() == 0 {
-		node.Text = n.Content(src)
-	}
-	for i := 0; i < int(n.ChildCount()); i++ {
-		child := n.Child(i)
-		if child == nil {
-			continue
-		}
-		node.Children = append(node.Children, convertNode(child, src))
-	}
-	return node
 }
